% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossmap_frac_in_genesets.R
\name{crossmap_frac_in_genesets}
\alias{crossmap_frac_in_genesets}
\title{Function to compute fraction of cross-mappable gene pairs}
\usage{
crossmap_frac_in_genesets(genes, crossmap_df, mappability_df,
  overlap_df = NULL, crossmap_threshold = 1e-06)
}
\arguments{
\item{genes}{character vector. Set of genes.}

\item{crossmap_df}{Cross-mappability data frame. First two columns should contain genes (character) and the 3rd column should contain cross-mappability score (numeric) from the 1st gene to the 2nd gene.}

\item{mappability_df}{Mappability data frame. First columns should contain genes (character) and the second columnd should contain corresponding mappability (numeric).}

\item{overlap_df}{Positional overlap data. First two columns should contain genes (character) with positional overlap.}

\item{crossmap_threshold}{numeric. A pair of genes are cross-mappable if cross-mappability score between them (in either direction) is greater or equal to this score.}
}
\value{
The fraction of cross-mappable gene pairs in a given gene set.
}
\description{
\code{crossmap_frac_in_genesets} computes fraction of cross-mappable gene pairs among all possible pairs of the given set of genes.
}
\details{
A pair of genes are called "cross-mappable" if cross-mappability score between them (in either direction) is greater or equal to a threshold score. 
If there are \code{n} genes in a set, there could have \code{n(n-1)/2} gene pairs. 
This function computes what fractio of all these pairs are cross-mappable.
Genes with mappability of NA are excluded from the gene set before computing the fraction,
as cross-mappability with those genes are undefined.
If positionally overlapped genes are provided (if overlap_df is not NULL), 
overlapped gene pairs are excluded from consideration.
}
