% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directed_to_undirected_crossmap.R
\name{directed_to_undirected_crossmap}
\alias{directed_to_undirected_crossmap}
\title{Undirected cross-mappability}
\usage{
directed_to_undirected_crossmap(crossmap_df, FUN = min)
}
\arguments{
\item{crossmap_df}{Directed cross-mappability data frame. First two columns should contain genes (character) and the 3rd column should contain cross-mappability score (numeric) from the 1st gene to the 2nd gene.}

\item{FUN}{function to summarize mappability scores from a vector of scores.}
}
\value{
A cross-mappability data frame.
}
\description{
\code{directed_to_undirected_crossmap} generates undirected cross-mappability scores from directed cross-mappability scores.
}
\details{
In a directed cross-mappability data frame,  cross-mappability score from gene A to gene B can be different from cross-mappability from gene B to gene A. This function generates an undirected version of cross-mappability score for each gene pair using a function (min, mean, max etc.).
}
