% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_crossmap.R
\name{filter_crossmap_by_genes}
\alias{filter_crossmap_by_genes}
\alias{filter_crossmap_by_score}
\alias{filter_crossmap_by_overlap}
\title{Filter cross-mappability data}
\usage{
filter_crossmap_by_genes(crossmap_df, incl.genes)

filter_crossmap_by_score(crossmap_df, min.crossmap = -Inf,
  max.crossmap = Inf)

filter_crossmap_by_overlap(crossmap_df, overlap_df)
}
\arguments{
\item{crossmap_df}{Cross-mappability data frame. First two columns should contain genes (character) and the 3rd column should contain cross-mappability score (numeric) from the 1st gene to the 2nd gene.}

\item{incl.genes}{character vector. Set of genes. Gene pairs within these genes are included in the filtered data.}

\item{min.crossmap}{numeric. Minimum cross-mappability score.}

\item{max.crossmap}{numeric. Maximum cross-mappability score.}

\item{overlap_df}{Position overlap data frame. First two columns should contain genes (character).}
}
\value{
A filttered cross-mappability data frame.
}
\description{
Filter cross-mappability data to include gene pairs where 1) each gene is within a given set of genes, 2) the cross-mappability score is within a given range, or 3) the genes are not positionally overlapped.
}
\details{
\code{filter_crossmap_by_genes} filters cross-mappability data to include pairs where each gene is within a given set of genes.
\code{filter_crossmap_by_score} filters cross-mappability data to include pairs where the cross-mappability score is within a given range.
\code{filter_crossmap_by_overlap} filters cross-mappability data to include pairs where the genes are not positionally overlapped.
}
